fix: PHPStan 0 errors, WPGraphQL test integration & CI pipeline
- Fix PHPStan errors: int cast on round() in DashboardRenderer,
  is_bool guard for WP_DEBUG_LOG in StatisticsProvider and
  SecurityDataProvider, updated phpstan.neon ignore patterns
- Add WPGraphQL test support: install script, bootstrap loader,
  adaptive test assertions for 4 GraphQL tests
- Wire WPGraphQL into run-quality-checks.sh and quality-checks.yml
  CI pipeline (FORCE_GRAPHQL_REINSTALL env var, chmod, reporting)
- Update CHANGELOG for v1.1.15 release
- Fix update-version-simple.sh: bash arrays for paths with spaces,
  --force flag for --no-confirm mode

+++ src/Admin/Renderer/DashboardRenderer.php
@@ -100,7 +100,7 @@ class DashboardRenderer {
 					<?php
 					RenderHelper::render_stat( $security_status['login_security']['max_attempts'], \__( 'Max Attempts', 'silver-assist-security' ) );
 					RenderHelper::render_stat( $security_status['overall']['blocked_ips_count'], \__( 'Blocked IPs', 'silver-assist-security' ), '', 'blocked-ips-card' );
-					RenderHelper::render_stat( round( $security_status['login_security']['lockout_duration'] / 60 ), \__( 'Lockout (min)', 'silver-assist-security' ) );
+					RenderHelper::render_stat( (int) round( $security_status['login_security']['lockout_duration'] / 60 ), \__( 'Lockout (min)', 'silver-assist-security' ) );
 					?>
 				</div>
 			</div>
